[pliron-llvm] Guard llvm-sys dependence under a default feature
