+++ run_ci.sh
@@ -51,3 +51,4 @@ if [ "$WASM_TESTS" = true ]; then
   cargo test --release --target wasm32-unknown-unknown
 fi
 RUSTDOCFLAGS="-D warnings" cargo doc --workspace
+RUSTFLAGS="-D warnings" cargo build -p pliron-llvm --no-default-features
